move topic mapping, reformat TODOs, get ready for brutal team review
