+++ beacon-chain/sync/BUILD.bazel
@@ -10,14 +10,12 @@ go_library(
         "rpc_hello.go",
         "service.go",
         "subscriber.go",
-        "topic_mapping.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/sync",
     visibility = ["//visibility:public"],
     deps = [
         "//beacon-chain/p2p:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
-        "//proto/eth/v1alpha1:go_default_library",
         "//shared:go_default_library",
         "//shared/roughtime:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
