+++ beacon-chain/sync/doc.go
@@ -1,4 +1,4 @@
 /*
-package sync ... TODO
+package sync ... TODO(3147): Add details on how sync works.
 */
 package sync
