+++ beacon-chain/sync/metrics.go
@@ -1,1 +1,3 @@
 package sync
+
+// TODO(3147): Add metrics for RPC & subscription success/error.
